saju-ai-agent/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.9
httpx>=0.27.0
PyJWT>=2.8.0
pytest>=8.0.0
